+++ zksync_sdk/ethereum_signer/web3.py
@@ -1,11 +1,9 @@
-from eth_account.messages import encode_defunct, defunct_hash_message
+from eth_account.messages import encode_defunct
 from eth_account.signers.base import BaseAccount
-from web3.auto import w3
-from typing import Optional
-from zksync_sdk.ethereum_signer.interface import EthereumSignerInterface, TxEthValidatorInterface
-from zksync_sdk.types import EncodedTx, SignatureType, TxEthSignature, Order
+from zksync_sdk.ethereum_signer.interface import EthereumSignerInterface
+from zksync_sdk.types import EncodedTx, SignatureType, TxEthSignature
 
-__all__ = ['EthereumSignerWeb3', 'TxEthValidator']
+__all__ = ['EthereumSignerWeb3']
 
 
 class EthereumSignerWeb3(EthereumSignerInterface):
@@ -22,24 +20,3 @@ class EthereumSignerWeb3(EthereumSignerInterface):
 
     def address(self) -> str:
         return self.account.address
-
-
-class TxEthValidator(TxEthValidatorInterface):
-
-    def is_valid_signature(self, tx: Order) -> bool:
-        msg = tx.human_readable_message().encode()
-
-        def get_sig(opt_value: Optional[TxEthSignature]) -> TxEthSignature:
-            if opt_value is None:
-                raise ValueError()
-            return opt_value
-
-        address = self.recover_signer_address(get_sig(tx.eth_signature), msg)
-        return self.signer_address == address
-
-    @staticmethod
-    def recover_signer_address(tx: TxEthSignature, message: bytes) -> str:
-        encoded_message = encode_defunct(message)
-        # INFO: remove prefix 0x
-        sig = bytes.fromhex(tx.signature[2:])
-        return w3.eth.account.recover_message(encoded_message, signature=sig)
