+++ zksync_sdk/types/transactions.py
@@ -6,7 +6,6 @@ from enum import Enum, IntEnum
 from typing import List, Optional, Union, Tuple
 
 from pydantic import BaseModel
-
 from zksync_sdk.lib import ZkSyncLibrary
 from zksync_sdk.serializers import (int_to_bytes, packed_amount_checked, packed_fee_checked,
                                     serialize_account_id,
@@ -530,6 +529,33 @@ class Order(EncodedTx):
             "signature": self.signature.dict() if self.signature else None,
             "ethSignature": self.eth_signature.dict() if self.eth_signature else None,
         }
+
+    def is_valid_eth_signature(self, signer_address: str) -> bool:
+        address = self._recover_signer_address()
+        return signer_address == address
+
+    def _recover_signer_address(self) -> str:
+        """
+        INFO: particular case implementation with dependency from Web3 interface
+              if it's needed to generelize for all Obejct type(Transfer, Swap etc) move to etherium_signer module
+              with interface & implemnetation for Web3 as Validator class( Visitor pattern )
+        """
+        from web3.auto import w3
+        from eth_account.messages import encode_defunct
+
+        msg = self.human_readable_message().encode()
+        encoded_message = encode_defunct(msg)
+
+        def get_sig(opt_value: Optional[TxEthSignature]) -> TxEthSignature:
+            if opt_value is None:
+                raise ValueError()
+            return opt_value
+
+        # INFO: remove prefix 0x
+        eth_sig = get_sig(self.eth_signature)
+        sig = bytes.fromhex(eth_sig.signature[2:])
+        return w3.eth.account.recover_message(encoded_message, signature=sig)
+
 
 
 @dataclass
